requirements.txt
dnspython
pygost
